TestTestgresCommon is updated
New tests:
 - test_users
 - test_poll_query_until
 - test_logging
